requirements.txt
discord.py
python-dotenv~=1.1.0
Quart~=0.20.0
hypercorn
PyGithub~=2.6.1
pip~=25.1
attrs~=25.3.0
cryptography~=45.0.2
supabase